refactor: changes from review
Various changes as a result of review feedback.
